+++ core/src/main/java/org/graphast/query/shortestpath/ShortestPathStrategy.java
@@ -4,7 +4,7 @@ import org.graphast.query.utils.DistanceVector;
 
 public interface ShortestPathStrategy {
 	
-	DistanceVector run(int sourceId);
-	DistanceVector run(int sourceId, int targetId);
+	DistanceVector run(long sourceId);
+	DistanceVector run(long sourceId, long targetId);
 
 }
