+++ core/src/main/java/org/graphast/query/utils/DistanceVector.java
@@ -4,47 +4,47 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 public class DistanceVector {
-	private HashMap<Integer, DistanceElement> vector;
+	private HashMap<Long, DistanceElement> vector;
 	
-	public DistanceVector(int sourceId) {
+	public DistanceVector(long sourceId) {
 		vector = new HashMap<>();
 		DistanceElement first = new DistanceElement(sourceId);
 		first.changeDistance(0);
 		vector.put(sourceId, first);
 	}
 
-	public DistanceElement getElement(int id) {
+	public DistanceElement getElement(long id) {
 		if (!vector.containsKey(id))
 			vector.put(id, new DistanceElement(id));
 		return vector.get(id);
 	}
 	
 	public void print() {
-		for (Integer n : vector.keySet()) {
+		for (Long n : vector.keySet()) {
 			DistanceElement element = vector.get(n);
 			System.out.println("Distance to node " + n + ": " + element.getDistance() + " | Previous node: " + element.getParentId());
 		}
 	}
 	
-	public void print(int sourceId, int targetId) {
+	public void print(long sourceId, long targetId) {
 		if (getElement(targetId).getParentId() == -1) {
 			System.out.println("No path between them");
 			return;
 		}
-		ArrayList<Integer> l = new ArrayList<>();
+		ArrayList<Long> l = new ArrayList<>();
 		l.add(targetId);
-		int parentId = targetId;
-		while ((parentId = vector.get(parentId).getParentId()) != -1) {
+		long parentId = targetId;
+		while ((parentId = vector.get(parentId).getParentId()) != -1l) {
 			l.add(0, parentId);
 			if (parentId == sourceId) break;
 		}
 		
-		for (Integer id : l)
+		for (Long id : l)
 			System.out.print(" -> " + id);
 		System.out.println();
 	}
 	
-	public double getDistance(int targetId) {
+	public double getDistance(long targetId) {
 		return vector.get(targetId).getDistance();
 	}
 }
